modules/chess-diagrams/requirements.txt
Flask
Pillow
gunicorn
